# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vets/search:
    post:
      operationId: searchVets
      summary: Find nearby clinics without persisting a visitor location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VetSearchInput"
      responses:
        "200":
          description: Distance-sorted nearby clinics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VetSearchResult"
        "400":
          description: Invalid search
        "429":
          description: Rate limited
        "503":
          description: Public map service unavailable
  /promotion/clinic-applications:
    post:
      operationId: createClinicApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicApplicationInput"
      responses:
        "201":
          description: Application saved, not published or emailed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionReceipt"
  /promotion/advertiser-inquiries:
    post:
      operationId: createAdvertiserInquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvertiserInquiryInput"
      responses:
        "201":
          description: Inquiry saved, no outbound messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionReceipt"
  /promotion/leads:
    get:
      operationId: getPromotionLeads
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Latest 100 leads of each kind, accessible only with the existing admin token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionLeads"
        "401":
          description: Unauthorized
components:
  schemas:
...
    VetSearchInput:
      type: object
      properties:
        query: { type: string, minLength: 2, maxLength: 120 }
        latitude: { type: number, minimum: -85, maximum: 85 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        urgent: { type: boolean, default: false }
        openNow: { type: boolean, default: false }
        emergencyOnly: { type: boolean, default: false }
    VetClinic:
      type: object
      required: [id, name, latitude, longitude, distanceKm, address, openingStatus, emergency, sponsored]
      properties:
        id: { type: string }
        name: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        distanceKm: { type: number }
        address: { type: string }
        phone: { type: string }
        website: { type: string }
        openingHours: { type: string }
        openingStatus: { type: string, enum: [open, closed, unknown] }
        emergency: { type: boolean }
        sponsored: { type: boolean }
    VetSearchResult:
      type: object
      required: [clinics, latitude, longitude, area, urgent, source, attribution, notice, radiusKm]
      properties:
        clinics:
          type: array
          items:
            $ref: "#/components/schemas/VetClinic"
        latitude: { type: number }
        longitude: { type: number }
        area: { type: string }
        urgent: { type: boolean }
        source: { type: string }
        attribution: { type: string }
        notice: { type: string }
        radiusKm: { type: number }
        mapTilesUrl: { type: string }
        mapAttribution: { type: string }
    ClinicApplicationInput:
      type: object
      required: [clinicName, contactEmail, city, message, consent]
      properties:
        clinicName: { type: string, minLength: 2, maxLength: 180 }
        contactName: { type: string, maxLength: 120 }
        contactEmail: { type: string, format: email, maxLength: 254 }
        city: { type: string, minLength: 2, maxLength: 120 }
        website: { type: string, maxLength: 2048 }
        message: { type: string, minLength: 10, maxLength: 4000 }
        intent: { type: string, enum: [claim, featured, both], default: claim }
        consent: { type: boolean }
        fax: { type: string, maxLength: 0 }
    AdvertiserInquiryInput:
      type: object
      required: [brandName, contactEmail, message, consent]
      properties:
        brandName: { type: string, minLength: 2, maxLength: 180 }
        contactName: { type: string, maxLength: 120 }
        contactEmail: { type: string, format: email, maxLength: 254 }
        website: { type: string, maxLength: 2048 }
        message: { type: string, minLength: 10, maxLength: 4000 }
        placements:
          type: array
          maxItems: 3
          items: { type: string, enum: [tools-guides, product-box, newsletter] }
        consent: { type: boolean }
        fax: { type: string, maxLength: 0 }
    PromotionReceipt:
      type: object
      required: [id, message]
      properties:
        id: { type: integer }
        message: { type: string }
    ClinicApplication:
      allOf:
        - $ref: "#/components/schemas/ClinicApplicationInput"
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: integer }
            createdAt: { type: string, format: date-time }
    AdvertiserInquiry:
      allOf:
        - $ref: "#/components/schemas/AdvertiserInquiryInput"
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: integer }
            createdAt: { type: string, format: date-time }
    PromotionLeads:
      type: object
      required: [clinicApplications, advertiserInquiries, clinicTotal, advertiserTotal]
      properties:
        clinicApplications:
          type: array
          items:
            $ref: "#/components/schemas/ClinicApplication"
        advertiserInquiries:
          type: array
          items:
            $ref: "#/components/schemas/AdvertiserInquiry"
        clinicTotal: { type: integer }
        advertiserTotal: { type: integer }
...