# lib/api-zod/src/generated/types/advertiserInquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdvertiserInquiryInput } from "./advertiserInquiryInput";

export type AdvertiserInquiry = AdvertiserInquiryInput & {
  id: number;
  createdAt: Date;
};