# lib/api-zod/src/generated/types/clinicApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClinicApplicationInput } from "./clinicApplicationInput";

export type ClinicApplication = ClinicApplicationInput & {
  id: number;
  createdAt: Date;
};