# lib/api-zod/src/generated/types/vetClinic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VetClinicOpeningStatus } from "./vetClinicOpeningStatus";

export interface VetClinic {
  id: string;
  name: string;
  latitude: number;
  longitude: number;
  distanceKm: number;
  address: string;
  phone?: string;
  website?: string;
  openingHours?: string;
  openingStatus: VetClinicOpeningStatus;
  emergency: boolean;
  sponsored: boolean;
}