# lib/api-zod/src/generated/types/vetSearchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VetSearchInput {
  /**
   * @minLength 2
   * @maxLength 120
   */
  query?: string;
  /**
   * @minimum -85
   * @maximum 85
   */
  latitude?: number;
  /**
   * @minimum -180
   * @maximum 180
   */
  longitude?: number;
  urgent?: boolean;
  openNow?: boolean;
  emergencyOnly?: boolean;
}